requirements.txt
networkx>=3.1
google-generativeai>=0.5.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
